Working model: bot check post - send to subscribe user - if a post has already been sent to a user, it will not be resent again;

+++ SpringDemoBot/src/main/java/io/proj3ct/SpringDemoBot/service/TelegramBot.java
@@ -4,8 +4,12 @@ import io.proj3ct.SpringDemoBot.config.BotConfig;
 import io.proj3ct.SpringDemoBot.model.User;
 import io.proj3ct.SpringDemoBot.model.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.tuple.Pair;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.methods.commands.SetMyCommands;
@@ -21,10 +25,10 @@ import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKe
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.KeyboardRow;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+import java.io.IOException;
 import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
+
 
 @Slf4j
 @Component
@@ -65,9 +69,160 @@ public class TelegramBot extends TelegramLongPollingBot {
     static final String UNSUBSCRIBE_QUESTION_TEXT = "Вы хотите отписаться от рассылки уведомлений?";
     static final String YES_BUTTON = "YES_BUTTON";
     static final String NO_BUTTON = "NO_BUTTON";
+    static final String CHOOSE_ROOM_COUNT = "Выберите количество комнат: ";
+    static final String ONE_ROOM_BUTTON = "ONE_ROOM_BUTTON";
+    static final String TWO_ROOM_BUTTON = "TWO_ROOM_BUTTON";
+
     static final String ERROR_TEXT = "Error occurred: ";
     static final String USER_ALREADY_REGISTERED_TEXT = "Эй! Вы уже зарегистрированы у нас. \uD83E\uDD16\uD83C\uDF1F";
     static final String NOT_SUBSCRIBED_WANT_TO_UNSUBSCRIBE_TEXT = "Эй! Вы еще не подписаны на наши уведомления. \uD83E\uDD16❌";
+
+
+
+    private List<Pair<String, String>> checkCian() {
+
+        List<Pair<String, String>> foundApartments = new ArrayList<>();
+
+        try {
+            Document doc = Jsoup.connect("https://kazan.cian.ru/snyat-kvartiru-1-komn-ili-2-komn/")
+                    .ignoreContentType(true)
+                    .userAgent("Mozilla/5.0 (Windows; U; Windows NT 6.1; rv:2.2) Gecko/20110201")
+                    .referrer("http://www.google.com")
+                    .timeout(12000)
+                    .followRedirects(true)
+                    .get();
+
+            // Проверка объявлений на длительный срок аренды и отправка уведомлений
+            String description = Objects.requireNonNull(doc.selectFirst("span._93444fe79c--color_black_100--Ephi7")).text();
+            String link = Objects.requireNonNull(doc.selectFirst("a._93444fe79c--link--VtWj6")).attr("href");
+            String date = Objects.requireNonNull(doc.selectFirst("div._93444fe79c--absolute--yut0v span._93444fe79c--text_letterSpacing__normal--tfToq")).text();
+
+            if (date.contains("сегодня")) {
+
+                foundApartments.add(Pair.of(description, link));
+                log.info("Успешно проверено и добавлено объявление: {}", description);
+                return foundApartments;
+            }
+
+
+        } catch (IOException e) {
+            log.error("Ошибка при попытке подключения к сайту CIAN: " + e.getMessage());
+        }
+
+        return Collections.emptyList();
+    }
+
+    private boolean checkApartments(List<Pair<String, String>> foundApartments) {
+
+        String apartmentType = "квартира";
+
+        for (Pair<String, String> apartment : foundApartments) {
+            String description = apartment.getLeft();
+            if (description.contains(apartmentType)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+
+    @Scheduled(fixedDelay = 6000)
+    private void executeCian() {
+
+        List<Pair<String, String>> foundApartments = checkCian();
+
+        if (!foundApartments.isEmpty() && checkApartments(foundApartments)) {
+            sendNotification(foundApartments);
+        }
+    }
+
+    private void sendNotification(List<Pair<String, String>> foundApartments) {
+
+
+        List<User> subscribedUsers = (List<User>) userRepository.findAll();
+
+        for (User user : subscribedUsers) {
+            Long chatId = user.getChatId();
+            SendMessage message = getSendMessage(foundApartments, user, chatId);
+            try {
+                if(message != null && !message.getChatId().isEmpty()) {
+                    execute(message);
+                    log.info("Уведомление успешно отправлено, id: {}", chatId);
+                }
+            } catch (TelegramApiException e) {
+                log.error("Ошибка при отправке уведомления: " + e.getMessage());
+            }
+        }
+        userRepository.saveAll(subscribedUsers);
+    }
+
+    private static SendMessage getSendMessage(List<Pair<String, String>> foundApartments, User user, Long chatId) {
+
+        String apartment_type = user.getApartmentType();
+
+        SendMessage message = new SendMessage();
+
+        for (Pair<String, String> apartment : foundApartments) {
+            String description = apartment.getLeft();
+            String link = apartment.getRight();
+
+            String lastPostedLink = user.getApartmentLink();
+
+            if(String.valueOf(description.charAt(0)).equals(apartment_type)) {
+                if(chatId != null && !Objects.equals(lastPostedLink, link)) {
+                    message.setChatId(String.valueOf(chatId));
+                    message.setText("Новая квартира: " + description + "\nСсылка: " + link);
+
+                    user.setApartmentLink(link);
+                    user.setApartmentDescription(description);
+
+                    return message;
+                }
+            } else {
+                return null;
+            }
+        }
+        return null;
+    }
+
+    private void setupApartmentMonitoringForm(Message msg) {
+
+        long chatId = msg.getChatId();
+
+        SendMessage message = new SendMessage();
+        message.setChatId(String.valueOf(chatId)); // Устанавливаем chatId
+        message.setText(CHOOSE_ROOM_COUNT); // Устанавливаем текст сообщения
+
+        InlineKeyboardMarkup markupInline = new InlineKeyboardMarkup();
+        List<List<InlineKeyboardButton>> rowsInline = getRoomCountKeyboardMarkup();
+        // Устанавливаем встроенную клавиатуру для сообщения
+        markupInline.setKeyboard(rowsInline);
+        message.setReplyMarkup(markupInline);
+        // Отправляем сообщение
+        executeMessage(message);
+    }
+
+    private static List<List<InlineKeyboardButton>> getRoomCountKeyboardMarkup() {
+        List<List<InlineKeyboardButton>> rowsInline = new ArrayList<>();
+        List<InlineKeyboardButton> rowInlineOne = new ArrayList<>();
+        // Создаем кнопку "1"
+        var yesButton = new InlineKeyboardButton();
+        yesButton.setText("1");
+        yesButton.setCallbackData(ONE_ROOM_BUTTON);
+        // Добавляем кнопки в строку
+        rowInlineOne.add(yesButton);
+
+        List<InlineKeyboardButton> rowInlineTwo = new ArrayList<>();
+        // Создаем кнопку "2"
+        var noButton = new InlineKeyboardButton();
+        noButton.setText("2");
+        noButton.setCallbackData(TWO_ROOM_BUTTON);
+        rowInlineTwo.add(noButton);
+        // Добавляем строку с кнопками
+        rowsInline.add(rowInlineOne);
+        rowsInline.add(rowInlineTwo);
+        return rowsInline;
+    }
 
     public TelegramBot(BotConfig config) {
         this.config = config;
@@ -79,6 +234,7 @@ public class TelegramBot extends TelegramLongPollingBot {
                 new BotCommand("/start", "Главное меню"),
                 new BotCommand("/subscribe", "Подписаться на рассылку"),
                 new BotCommand("/delete", "Отписаться от уведомлений и удалить свои данные"),
+                new BotCommand("/cian", "функция Циан"),
                 new BotCommand("/help", "Помощь по управлению ботом")
         );
         try {
@@ -101,7 +257,11 @@ public class TelegramBot extends TelegramLongPollingBot {
     @Override
     public void onUpdateReceived(Update update) {
         if (isTextMessage(update)) {
-            processTextMessage(update);
+            try {
+                processTextMessage(update);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         } else if (isCallbackQuery(update)) {
             processCallbackQuery(update);
         }
@@ -115,7 +275,7 @@ public class TelegramBot extends TelegramLongPollingBot {
         return update.hasCallbackQuery();
     }
 
-    private void processTextMessage(Update update) {
+    private void processTextMessage(Update update) throws IOException {
         String messageText = update.getMessage().getText();
         long chatId = update.getMessage().getChatId();
 
@@ -134,7 +294,7 @@ public class TelegramBot extends TelegramLongPollingBot {
         }
     }
 
-    private void processUserCommand(String messageText, long chatId, Update update) {
+    private void processUserCommand(String messageText, long chatId, Update update) throws IOException {
         switch (messageText) {
             case "/start":
                 startCommandReceived(chatId, update.getMessage().getChat().getFirstName());
@@ -144,6 +304,9 @@ public class TelegramBot extends TelegramLongPollingBot {
                 break;
             case "/delete", "Отписаться":
                 delete(update.getMessage());
+                break;
+            case "/cian":
+                setupApartmentMonitoringForm(update.getMessage());
                 break;
             case "Вывести мои данные":
                 showMyData(update.getMessage());
@@ -162,10 +325,10 @@ public class TelegramBot extends TelegramLongPollingBot {
         long chatId = update.getCallbackQuery().getMessage().getChatId();
         String messageText = update.getCallbackQuery().getMessage().getText();
 
-        if (callbackData.equals(YES_BUTTON)) {
-            processYesButton(update, messageText);
-        } else if (callbackData.equals(NO_BUTTON)) {
-            processNoButton(messageText, chatId, messageId);
+        switch (callbackData) {
+            case YES_BUTTON -> processYesButton(update, messageText);
+            case NO_BUTTON -> processNoButton(messageText, chatId, messageId);
+            case ONE_ROOM_BUTTON, TWO_ROOM_BUTTON -> processRoomCountButton(update);
         }
     }
 
@@ -184,6 +347,27 @@ public class TelegramBot extends TelegramLongPollingBot {
             executeEditMessageText(SUBSCRIBE_FALSE_TEXT, chatId, messageId);
         } else if (messageText.equals(UNSUBSCRIBE_QUESTION_TEXT)) {
             executeEditMessageText(UNSUBSCRIBE_FALSE_TEXT, chatId, messageId);
+        }
+    }
+
+    private void processRoomCountButton(Update update) {
+
+        Optional<User> optionalUser = userRepository.findById(update.getCallbackQuery().getMessage().getChatId());
+
+        if(optionalUser.isPresent()) {
+            User user = optionalUser.get();
+            try {
+                if (update.getCallbackQuery().getData().contains(ONE_ROOM_BUTTON)) {
+                    user.setApartmentType("1");
+                } else if (update.getCallbackQuery().getData().contains(TWO_ROOM_BUTTON)) {
+                    user.setApartmentType("2");
+                }
+                executeEditMessageText("Форма заполнена!", update.getCallbackQuery().getMessage().getChatId(), update.getCallbackQuery().getMessage().getMessageId());
+                userRepository.save(user);
+                log.info("Apartment's type saved: " + user);
+            } catch (DataAccessException e) {
+                log.error("Error saving apartment's type for user: {}", e.getMessage());
+            }
         }
     }
 
@@ -206,9 +390,10 @@ public class TelegramBot extends TelegramLongPollingBot {
             userRepository.save(user);
             executeEditMessageText(SUBSCRIBE_TRUE_TEXT, chatId, msg.getMessageId());
             log.info("User saved: {}", user);
+
+            setupApartmentMonitoringForm(msg);
         } catch (DataAccessException e) {
             log.error("Error saving user to the database: {}", e.getMessage());
-            // Обработка ошибки сохранения в базу данных
         }
     }
 
@@ -223,8 +408,9 @@ public class TelegramBot extends TelegramLongPollingBot {
             message.setChatId(String.valueOf(chatId)); // Устанавливаем chatId
             message.setText("Вы хотите подписаться на рассылку уведомлений?"); // Устанавливаем текст сообщения
 
+
             // Создаем встроенную клавиатуру
-            InlineKeyboardMarkup markupInline = getKeyboardMarkup();
+            InlineKeyboardMarkup markupInline = getYesNoKeyboardMarkup();
             message.setReplyMarkup(markupInline);
             // Отправляем сообщение
             executeMessage(message);
@@ -234,7 +420,7 @@ public class TelegramBot extends TelegramLongPollingBot {
         }
     }
 
-    private static InlineKeyboardMarkup getKeyboardMarkup() {
+    private static InlineKeyboardMarkup getYesNoKeyboardMarkup() {
         InlineKeyboardMarkup markupInline = new InlineKeyboardMarkup();
         List<List<InlineKeyboardButton>> rowsInline = new ArrayList<>();
         List<InlineKeyboardButton> rowInline = new ArrayList<>();
@@ -260,14 +446,16 @@ public class TelegramBot extends TelegramLongPollingBot {
 
         Optional<User> optionalUser = userRepository.findById(msg.getChatId());
         // Если пользователь с таким chatId найден, выполняем удаление пользователя из базы данных
-        User user = optionalUser.get();
-        // Удаляем пользователя из базы данных
-        try {
-            userRepository.delete(user);
-            executeEditMessageText(UNSUBSCRIBE_TRUE_TEXT, msg.getChatId(), msg.getMessageId());
-            log.info("User deleted: " + user);
-        } catch (DataAccessException e) {
-            log.error("Error delete user from the database: {}", e.getMessage());
+        if(optionalUser.isPresent()) {
+            User user = optionalUser.get();
+            // Удаляем пользователя из базы данных
+            try {
+                userRepository.delete(user);
+                executeEditMessageText(UNSUBSCRIBE_TRUE_TEXT, msg.getChatId(), msg.getMessageId());
+                log.info("User deleted: " + user);
+            } catch (DataAccessException e) {
+                log.error("Error delete user from the database: {}", e.getMessage());
+            }
         }
     }
 
@@ -282,7 +470,7 @@ public class TelegramBot extends TelegramLongPollingBot {
 
             // Создаем встроенную клавиатуру
 
-            InlineKeyboardMarkup markupInline = getInlineKeyboardMarkup();
+            InlineKeyboardMarkup markupInline = getYesNoKeyboardMarkup();
             message.setReplyMarkup(markupInline);
             // Отправляем сообщение
             executeMessage(message);
@@ -290,10 +478,6 @@ public class TelegramBot extends TelegramLongPollingBot {
             prepareAndSendMessage(msg.getChatId(), NOT_SUBSCRIBED_WANT_TO_UNSUBSCRIBE_TEXT);
             log.warn("User not found for chatId: " + chatId);
         }
-    }
-
-    private static InlineKeyboardMarkup getInlineKeyboardMarkup() {
-        return getKeyboardMarkup();
     }
 
     private void showMyData(Message msg) {
